WebRTC Transport Refactor (#4502)
* start refactoring server transport

* remove unused transport stuff

* remove a bunch more unused code

* remove even more stuff

* fix type for routers

* fix some media stuff

* fix tests
