+++ packages/engine/src/initializeEngine.ts
@@ -87,10 +87,6 @@ const configureEditor = async (options: Required<InitializeOptions>) => {
 const configureServer = async (options: Required<InitializeOptions>, isMediaServer = false) => {
   Engine.scene = new Scene()
 
-  // Had to add this to make mocha tests pass
-  Network.instance ||= new Network()
-  Network.instance.isInitialized = true
-
   EngineEvents.instance.once(EngineEvents.EVENTS.JOINED_WORLD, () => {
     console.log('joined world')
     dispatchLocal(EngineActions.enableScene({ renderer: true, physics: true }) as any)
@@ -304,7 +300,6 @@ export const initializeEngine = async (initOptions: InitializeOptions = {}): Pro
 
   if (options.type === EngineSystemPresets.CLIENT) {
     EngineEvents.instance.once(EngineEvents.EVENTS.CONNECT, ({ id }) => {
-      Network.instance.isInitialized = true
       Engine.userId = id
     })
   } else if (options.type === EngineSystemPresets.SERVER) {
