--- packages/engine/src/networking/interfaces/NetworkSchema.ts
@@ -1,9 +1,0 @@
-/** Interface for Network schema. */
-export interface NetworkSchema {
-  /** Transporter of the message. */
-  transport: any
-  /** List of supported message types. */
-  messageTypes: {
-    [key: string]: any
-  }
-}
