--- packages/engine/src/networking/templates/DefaultNetworkSchema.ts
@@ -1,9 +1,0 @@
-import { MessageTypes } from '../enums/MessageTypes'
-import { NetworkSchema } from '../interfaces/NetworkSchema'
-
-export const DefaultNetworkSchema: NetworkSchema = {
-  transport: null,
-  messageTypes: {
-    ...MessageTypes
-  }
-}
