+++ packages/gameserver/src/NetworkFunctions.ts
@@ -1,23 +1,18 @@
 import { Engine } from '@xrengine/engine/src/ecs/classes/Engine'
-import { EngineEvents } from '@xrengine/engine/src/ecs/classes/EngineEvents'
-import { getComponent, hasComponent } from '@xrengine/engine/src/ecs/functions/ComponentFunctions'
+import { getComponent } from '@xrengine/engine/src/ecs/functions/ComponentFunctions'
 import { Network } from '@xrengine/engine/src/networking/classes/Network'
 import { MessageTypes } from '@xrengine/engine/src/networking/enums/MessageTypes'
 import { DataConsumer, DataProducer } from 'mediasoup/node/lib/types'
 import logger from '@xrengine/server-core/src/logger'
 import config from '@xrengine/server-core/src/appconfig'
 import { closeTransport } from './WebRTCFunctions'
-import { Quaternion, Vector3 } from 'three'
 import { SpawnPoints } from '@xrengine/engine/src/avatar/AvatarSpawnSystem'
 import { NetworkObjectComponent } from '@xrengine/engine/src/networking/components/NetworkObjectComponent'
 import { NetworkWorldAction } from '../../engine/src/networking/functions/NetworkWorldAction'
-import { XRInputSourceComponent } from '../../engine/src/xr/components/XRInputSourceComponent'
 import { Action } from '@xrengine/engine/src/networking/interfaces/Action'
 import { dispatchFrom } from '@xrengine/engine/src/networking/functions/dispatchFrom'
 import { UserId } from '@xrengine/common/src/interfaces/UserId'
-import { XRHandsInputComponent } from '@xrengine/engine/src/xr/components/XRHandsInputComponent'
 import { SocketWebRTCServerTransport } from './SocketWebRTCServerTransport'
-import { AvatarComponent } from '@xrengine/engine/src/avatar/components/AvatarComponent'
 
 const gsNameRegex = /gameserver-([a-zA-Z0-9]{5}-[a-zA-Z0-9]{5})/
 
@@ -127,79 +122,6 @@ export function getUserIdFromSocketId(socketId) {
   return client?.userId
 }
 
-export async function validateNetworkObjects(): Promise<void> {
-  if (!Engine.isInitialized) return
-  const world = Engine.currentWorld!
-  for (const [userId, client] of world.clients) {
-    // Validate that user has phoned home recently
-    if (process.env.APP_ENV !== 'development' && Date.now() - client.lastSeenTs > 30000) {
-      console.log('Removing client ', userId, ' due to inactivity')
-      if (!client) return console.warn('Client is not in client list')
-
-      const disconnectedClient = Object.assign({}, client)
-
-      dispatchFrom(world.hostId, () => NetworkWorldAction.destroyClient({ $from: userId }))
-
-      console.log('Disconnected Client:', disconnectedClient.userId)
-      if (disconnectedClient?.instanceRecvTransport) {
-        console.log('Closing instanceRecvTransport')
-        await disconnectedClient.instanceRecvTransport.close()
-        console.log('Closed instanceRecvTransport')
-      }
-      if (disconnectedClient?.instanceSendTransport) {
-        console.log('Closing instanceSendTransport')
-        await disconnectedClient.instanceSendTransport.close()
-        console.log('Closed instanceSendTransport')
-      }
-      if (disconnectedClient?.channelRecvTransport) {
-        console.log('Closing channelRecvTransport')
-        await disconnectedClient.channelRecvTransport.close()
-        console.log('Closed channelRecvTransport')
-      }
-      if (disconnectedClient?.channelSendTransport) {
-        console.log('Closing channelSendTransport')
-        await disconnectedClient.channelSendTransport.close()
-        console.log('Closed channelSendTransport')
-      }
-
-      console.log('Removed transports for', userId)
-
-      // Find all network objects that the disconnecting client owns and remove them
-      for (const eid of world.getOwnedNetworkObjects(userId)) {
-        const { networkId } = getComponent(eid, NetworkObjectComponent)
-        dispatchFrom(world.hostId, () => NetworkWorldAction.destroyObject({ networkId }))
-      }
-      if (world.clients.has(userId)) world.clients.delete(userId)
-      console.log('Finished removing inactive client', userId)
-    }
-  }
-  /*
-  Object.keys(Network.instance.networkObjects).forEach((key: string) => {
-    const networkObject = Network.instance.networkObjects[key]
-    // Validate that the object has an associated user and doesn't belong to a non-existant user
-    if (
-      !hasComponent(networkObject.entity, AvatarComponent) ||
-      (networkObject.uniqueId !== undefined && world.clients.get(networkObject.uniqueId) !== undefined)
-    )
-      return
-
-    console.log('Culling ownerless object: ', key, 'owned by ', networkObject.uniqueId)
-
-    // If it does, tell clients to destroy it
-    dispatchFromServer(NetworkWorldAction.destroyObject(Number(key)))
-
-    // get network object
-    const entity = networkObject.entity
-
-    // Remove the entity and all of it's components
-    removeEntity(entity)
-
-    // Remove network object from list
-    delete Network.instance.networkObjects[key]
-    logger.info(key, ' removed from simulation')
-  })*/
-}
-
 export async function handleConnectToWorld(socket, data, callback, userId: UserId, user, avatarDetail): Promise<any> {
   const transport = Network.instance.transport as SocketWebRTCServerTransport
 
@@ -302,10 +224,6 @@ export function handleIncomingActions(socket, message) {
     world.outgoingActions.add(a)
   }
   // console.log('SERVER INCOMING ACTIONS', JSON.stringify(actions))
-}
-
-export async function handleIncomingMessage(socket, message): Promise<any> {
-  Network.instance.incomingMessageQueueReliable.add(message)
 }
 
 export async function handleHeartbeat(socket): Promise<any> {
