+++ packages/gameserver/src/app.ts
@@ -103,6 +103,7 @@ export const createApp = (): Application => {
             }
           },
           (io) => {
+            Network.instance = new Network()
             Network.instance.transport = new SocketWebRTCServerTransport(app)
             Network.instance.transport.initialize()
             io.use((socket, next) => {
